% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggpar.R
\name{aggpar}
\alias{aggpar}
\title{Aggregate Parameters for GTAP Data}
\usage{
aggpar(mapfile, setfile, datfile, stdprm, dir)
}
\arguments{
\item{mapfile}{Character string. Path to the mapping file for aggregation.}

\item{setfile}{Character string. Path to the set file for disaggregated GTAP data.}

\item{datfile}{Character string. Path to the disaggregated GTAP data file.}

\item{stdprm}{Character string. Path to the standard parameters file.}

\item{dir}{Character string. Directory where aggregated output files will be saved.}
}
\value{
This function does not return a value. It creates aggregated parameter files in the specified directory.
}
\description{
This function aggregates parameters for GTAP (Global Trade Analysis Project) data.
It performs consistency checks, creates necessary configuration files, and executes
the parameter aggregation process.
}
\details{
The function performs the following steps:
\enumerate{
\item Runs \code{aggcheck()} to create gmap.har and check consistency between input sets and data.
\item Creates the output directory if it doesn't exist.
\item Generates an 'aggpar.cmf' configuration file with input and output file specifications.
\item Executes the 'aggpar' system command to perform the actual parameter aggregation.
}

Input files:
\itemize{
\item datfile: Disaggregated data in GTAP notation
\item stdprm: Disaggregated parameters
\item gmap.har: User-supplied parameters from input file (created by aggcheck)
\item setfile: Set specification for disaggregate data
}

Output files:
\itemize{
\item aggpar.har: Aggregated CDE, CES, and CET elasticities (saved in the specified directory)
}
}
\examples{
\dontrun{
aggpar("path/to/mapfile", "path/to/setfile", "path/to/datfile",
       "path/to/stdprm", "output_directory")
}

}
\seealso{
\code{\link{aggcheck}}
}
