% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files_to_dir.r
\name{copy_files_to_dir}
\alias{copy_files_to_dir}
\title{Copy .tab Files to a Destination Directory}
\usage{
copy_files_to_dir(dest_folder = "gempack", extension = "tab")
}
\arguments{
\item{dest_folder}{The name of the destination folder. Defaults to "gempack".
The folder will be created inside the current working directory (getwd()).}

\item{extension}{The file extension to copy. Defaults to "tab".}
}
\value{
Invisibly returns a character vector of the full paths of the
successfully copied files.  Also prints a message to the console
indicating the number of files copied.
}
\description{
This function copies .tab files from the package's 'extdata/tablo' directory
to a specified destination directory.
}
