% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_missing_tab_files.r
\name{copy_missing_tab_files}
\alias{copy_missing_tab_files}
\title{Copy Missing .tab Files to a Destination Directory}
\usage{
copy_missing_tab_files(dest_folder = "gempack", missing_exe_files)
}
\arguments{
\item{dest_folder}{The name of the destination folder. Defaults to "gempack".
The folder will be created inside the current working directory (getwd()).}

\item{missing_exe_files}{A character vector of missing .exe filenames.}
}
\value{
Invisibly returns a character vector of the full paths of the
successfully copied .tab files. Also prints messages to the console
indicating the number of files copied and suggesting a command to produce .exe files.
}
\description{
This function copies missing .tab files from the package's 'extdata/tablo' directory
to a specified destination directory, based on a list of missing .exe files.
}
\details{
The function converts the provided .exe filenames to .tab filenames and attempts to copy
the corresponding .tab files. It will only copy files that exist in the package's
'extdata/tablo' directory.
}
