% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createdat.R
\name{createdat}
\alias{createdat}
\title{Create Aggregated Data and Parameters for GTAP}
\usage{
createdat(mapfile, setfile, datfile, stdprm, dir)
}
\arguments{
\item{mapfile}{Gempack text file with aggregation template,
generated by gtapshape.}

\item{setfile}{HAR file with GTAP sets, including the subnational
bounds, also generated by gtapshape.}

\item{datfile}{HAR file with GTAP data, including the land use dat
by subnational bounds, also generated by gtapshape.}

\item{stdprm}{HAR file with parameters for the standard GTAP
database..}

\item{dir}{Character string. Directory where output files will be
saved.}
}
\value{
This function does not return a value. It creates output files in the specified directory.
}
\description{
This function performs a series of operations to create aggregated data and parameters
for GTAP (Global Trade Analysis Project) modeling. It aggregates data, processes land use
parameters, generates a GTAP view, and packages the resulting files.
}
\details{
The function performs the following steps:
\enumerate{
\item Aggregates data using \code{aggdat()}.
\item Processes land use parameters using \code{lusepar()}.
\item Generates a GTAP view using \code{gtapview()}.
\item Creates a zip file containing input files.
\item Removes temporary files.
}
}
\examples{
\dontrun{
createdat("path/to/mapfile", "path/to/setfile", "path/to/datfile",
          "path/to/stdprm", "output_directory")
}

}
\seealso{
\code{\link{aggdat}}, \code{\link{lusepar}}, \code{\link{gtapview}}
}
