% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtapview.R
\name{gtapview}
\alias{gtapview}
\title{Generate GTAP View Files}
\usage{
gtapview(dir)
}
\arguments{
\item{dir}{Character string. Directory containing input files and where output files will be saved.}
}
\value{
This function does not return a value. It creates view files in the specified directory.
}
\description{
This function creates configuration files and executes the 'gtapview' command
to generate view files for GTAP (Global Trade Analysis Project) data.
}
\details{
The function performs the following steps:
\enumerate{
\item Creates a 'gtapview.cmf' configuration file specifying input and output files.
\item Executes the 'gtapview' system command to generate view files.
}

Input files (expected in the specified directory):
\itemize{
\item basedata.har: Aggregated GTAP data
\item sets.har: Aggregated GTAP sets
\item default.prm: Aggregated data parameters
}

Output files (created in the specified directory):
\itemize{
\item baseview.har: File with headers for viewing
\item baserate.har: File with tax rate headers for viewing
}
}
\examples{
\dontrun{
gtapview("path/to/data_directory")
}

}
