% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lusepar.R
\name{lusepar}
\alias{lusepar}
\title{Process Land Use Parameters for GTAP}
\usage{
lusepar(mapfile, setfile, datfile, stdprm, dir)
}
\arguments{
\item{mapfile}{Gempack text file with aggregation template,
generated by gtapshape.}

\item{setfile}{HAR file with GTAP sets, including the subnational
bounds, also generated by gtapshape.}

\item{datfile}{HAR file with GTAP data, including the land use dat
by subnational bounds, also generated by gtapshape.}

\item{stdprm}{HAR file with parameters for the standard GTAP
database..}

\item{dir}{Character string. Directory where output files will be
saved.}
}
\value{
This function does not return a value. It creates and modifies files in the specified directory.
}
\description{
Aggregates parameters of the GTAP database.
}
\details{
The function performs the following steps:
\enumerate{
\item Calls \code{aggpar()} to aggregate parameters.
\item Creates a 'lusepar.cmf' configuration file specifying input and output files.
\item Executes the 'lusepar' system command to process land use parameters.
\item Removes the temporary 'aggpar.har' file.
}
}
\examples{
\dontrun{
lusepar("path/to/mapfile", "path/to/setfile", "path/to/datfile",
        "path/to/stdprm", "output_directory")
}

}
\seealso{
\code{\link{aggpar}}
}
