% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptext2har.R
\name{maptext2har}
\alias{maptext2har}
\title{Convert GTAP Aggregation Template to HAR File}
\usage{
maptext2har(mapfile)
}
\arguments{
\item{mapfile}{An aggregation template for the GTAP database.}
}
\value{
This function does not return a value. It writes the extracted data to a HAR file named 'test.har'.
}
\description{
This function reads a GTAP aggregation template file and converts it into a HAR (Header Array) file.
}
\details{
The function performs the following steps:
\enumerate{
\item Extracts various sections from the GTAP aggregation template using the extract.values function.
\item Creates a structured list (aggregation) containing the extracted data.
\item Uses the HARr package to write the structured data to a HAR file.
}
}
\note{
\itemize{
\item This function requires the HARr package to be installed.
\item The extract.values function should be defined elsewhere in the package.
\item The output HAR file is always named 'test.har' and is written to the current working directory.
}
}
\examples{
\dontrun{
maptext2har("path/to/gtap_aggregation_template.txt")
}

}
