% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gtap_program.R
\name{run_gtap_program}
\alias{run_gtap_program}
\title{Run a GEMPACK Program}
\usage{
run_gtap_program(program, cmf)
}
\arguments{
\item{program}{The name of the GEMPACK program to run (e.g., "tablo").  This should be the full path of the executable or be in the system's PATH.}

\item{cmf}{The path to the command file (.cmf) that controls the GEMPACK program.}
}
\value{
Depends on the program, but the message should be that the
program completed without error.
}
\description{
This function executes a GEMPACK program using the system command.
}
