% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitlr2.R
\name{splitlr2}
\alias{splitlr2}
\title{Split Land Rent and Generate GTAP Land Use Data}
\usage{
splitlr2(landdat = NULL, landsets = NULL, stdgtapsets, stdgtapdata, dir)
}
\arguments{
\item{landdat}{Character string. Path to the land use and cover
data file.}

\item{landsets}{Character string. Path to the sets file for land
use data.}

\item{stdgtapsets}{Character string. Path to the sets and mapping
data file for Standard GTAP v.11c.}

\item{stdgtapdata}{Character string. Path to the Standard GTAP
v.11c data file.}

\item{dir}{Character string. Directory name for output files.}
}
\value{
No return value. Files are written to the specified directory.
}
\description{
This function creates a GEMPACK command file (CMF) and executes a GEMPACK program
to split land rent and generate GTAP land use data.
}
\details{
This function performs the following steps:
\enumerate{
\item Creates the output directory if it doesn't exist.
\item Generates a GEMPACK command file (splitlr.cmf) with file paths and settings.
\item Executes the GEMPACK program '03_rentcalc' using the generated CMF file.
\item Moves the CMF file to the output directory.
}
}
\examples{
\dontrun{
splitlr("landdata.har", "landsets.har", "gtapsets.har", "basedata.har", "forestdata.har", "output_dir")
}

}
