% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_gtapdata.R
\name{unzip_gtapdata}
\alias{unzip_gtapdata}
\title{Unzip GTAP Data File}
\usage{
unzip_gtapdata()
}
\value{
This function does not return a value. It unzips the file and prints a success
message if the operation is successful. If the unzip operation fails, it stops
with an error message.
}
\description{
This function unzips a password-protected GTAP data file from the gtapshapeagg package.
It prompts the user for a password and uses 7-Zip to extract the contents.
}
\details{
The function locates the zip file 'gsdgdat11cMV6.zip' in the 'GTAPv11c' directory
of the gtapshapeagg package. It then prompts the user for a password and uses
7-Zip to extract the contents to the same directory.
}
\note{
This function requires 7-Zip to be installed and accessible from the command line.
}
\examples{
\dontrun{
unzip_gtapdata()
}

}
